Исправил запрос для вывода статистики

+++ src/main/java/search_engine/dto/statistics/TotalStatistics.java
@@ -6,8 +6,8 @@ import lombok.experimental.Accessors;
 @Data
 @Accessors(chain = true)
 public class TotalStatistics {
-    private int sites;
-    private int pages;
-    private int lemmas;
+    private long sites;
+    private long pages;
+    private long lemmas;
     private boolean indexing;
 }
